feat : formatData 함수 구현
