+++ src/app/mypage/questions/page.tsx
@@ -11,6 +11,7 @@ import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import { ROUTE_PATHS } from '@/constants/config';
 import { getQuestionApi } from '@/services/question';
+import { formatDate } from '@/utils/fomatDate';
 
 interface Choice {
   content: string;
@@ -46,14 +47,6 @@ export default function MyQuestions() {
   }, []);
 
   const handleClickButton = () => {};
-  function formatDate(dateString: string) {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-
-    return `${year}.${month}.${day}`;
-  }
 
   return (
     <Wrapper>
